[test] 불필요한 테스트 제거
